@@ -1,9 +1,10 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
-using System.Web.Mvc;
 using System.Web.SessionState;
+using Microsoft.AspNetCore.Mvc;
+
 
 namespace Bookstore.Web.Helpers
 {
